requirements.txt
pynput>=1.7.6
pygetwindow>=0.0.9
pyautogui>=0.9.54
requests>=2.31.0
PySocks>=1.7.1

